+++ scripts/deepl.ts
@@ -5,6 +5,7 @@ import fs from "node:fs"
 import path from "node:path"
 
 config()
+console.log("Imported env variables:", process.env.DEEPL_API_KEY ? "DEEPL_API_KEY: ✅" : "DEEPL_API_KEY; ❌")
 // set DEEPL_API_KEY=... in terminal
 if (process.argv.length === 2) {
 	exec(
@@ -27,29 +28,26 @@ if (process.argv.length === 3) {
 	console.log("Update the language files")
 	// read the deepl.Lang.json files and copy the new keys into the language files in the folder firefox _locales
 
-	const localesPath = "src\\locales"
+	const localesPath = path.join("src", "locales")
+	const translationPath = ".translation"
 
 	// Step 1: Read the deepl.Lang.json file
 	const deepLang = ["DE", "EN", "ES", "FR", "IT", "JA", "KO", "PL", "PT", "PT", "SV", "TR", "ZH"]
 	const firefoxLocales = ["de", "en", "es", "fr", "it", "ja", "ko", "pl", "pt", "pt_BR", "sv", "tr", "zh_CN"]
 	deepLang.forEach((trans, index) => {
-		fs.readFile(path.join(".translation\\", "deepl." + trans + ".json"), "utf8", function (err, newKeys) {
+		fs.readFile(path.join(translationPath, `deepl.${trans}.json`), "utf8", function (err, newKeys) {
 			console.log("lang:", trans)
 			if (err) return console.log(err)
 
 			// Append the new keys to the language files
 			console.log(firefoxLocales[index])
-			fs.readFile(localesPath + "\\" + firefoxLocales[index] + ".json", "utf8", function (err, oldKeys) {
+			const localeFilePath = path.join(localesPath, `${firefoxLocales[index]}.json`)
+			fs.readFile(localeFilePath, "utf8", function (err, oldKeys) {
 				if (err) return console.log(err)
 				const newTranslation = { ...JSON.parse(oldKeys), ...JSON.parse(newKeys) }
-				fs.writeFile(
-					localesPath + "\\" + firefoxLocales[index] + ".json",
-					JSON.stringify(newTranslation, null, 2),
-					"utf8",
-					function (err) {
-						if (err) return console.log(err)
-					},
-				)
+				fs.writeFile(localeFilePath, JSON.stringify(newTranslation, null, 2), "utf8", function (err) {
+					if (err) return console.log(err)
+				})
 			})
 		})
 	})